fastapi/app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-docx==1.1.0
Pillow==10.2.0
lxml==5.1.0
python-multipart==0.0.6
officemath2latex==0.1.0
latex2mathml==3.77.0
pydantic==2.5.3
mysql-connector-python==9.0.0
